move new menu to menu module
